recipes/model-customization/finetune-automation/requirements.txt
streamlit
requests
boto3
ruamel.yaml
pyyaml
sagemaker
perplexity
json
os
time
sys
tarfile
io
json
argparse
datetime
